refactor: move shadow functionality
