+++ packages/eui-theme-common/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@types/lodash": "^4.14.202",
+    "chroma-js": "^2.4.2",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
@@ -32,6 +33,7 @@
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.21.5",
     "@emotion/react": "^11.11.0",
+    "@types/chroma-js": "^2.4.0",
     "@types/jest": "^29.5.12",
     "@types/prettier": "2.7.3",
     "@types/react": "^16.9 || ^17.0 || ^18.0",
