+++ packages/eui/src/global_styling/mixins/index.ts
@@ -13,3 +13,4 @@ export * from './_states';
 export * from './_typography';
 export * from './_responsive';
 export * from './_button';
+export * from './_shadow';
